Replace VERSION_STRING with the STRING_GIT_VERSION
After talking to EvdZ it became clear that the version string is not a
reliable source and just an unused default.

When STRING_GIT_VERSION is set, the VERSION_STRING will be filled with
that instead. Additionally, the VERSION_STRING is now added to the M115
command, so the version also shows on a M115, not just on startup.

Signed-off-by: Olliver Schinagl <o.schinagl@ultimaker.com>